Upgrade Gradle to 6.8.1
Due to the upgrade to the latest Gradle version 6.8.1 the configuration can be
reduced by replacing custom tasks with build-in tasks for javadoc and source
jar file generation. Additionally the build entvironment can be now set to use
Java 8 by default because of the introduction of the toolchain support. This
gurantees that all builds are using the same Java version.
Additonally the jfrog plugins have been removed as because Bintray/JCenter are
going offline on 1st of May. see https://jfrog.com/blog/into-the-sunset-bintray-jcenter-gocenter-and-chartcenter/