preserve order of classpath entries when detecting which classes are used from which dependency
this resolves an issue when two classes are present in different dependencies with same name and package.
Default behaviour from java is to check each jar on classpath in order to find the correct one (the one that should be used).
still this can have some side effects, but at least the behaviour is consistent.